+++ src/widgets/TbDataColumn.php
@@ -8,6 +8,7 @@
  */
 
 Yii::import('zii.widgets.grid.CDataColumn');
+Yii::import('booster.helpers.TbIcon');
 
 /**
  *## Bootstrap grid data column.
@@ -38,7 +39,7 @@ class TbDataColumn extends CDataColumn
 
 
 			if ($sort->resolveAttribute($this->name) !== false)
-				$label .= ' <span class="caret"></span>';
+				$label .= ' ' . TbIcon::render('bi-caret-down-fill');
 
 			echo $sort->link($this->name, $label, array('class' => 'sort-link'));
 		} else {